Исправлено получение списка заказов

+++ internal/models/order.go
@@ -3,8 +3,8 @@ package models
 import "time"
 
 type Order struct {
-	ID        int       `json:id`
-	UserID    int       `json:user_id`
+	ID        int64     `json:id`
+	UserID    int64     `json:user_id`
 	Number    string    `json:number`
 	CreatedAt time.Time `json:created_at`
 }
